!13 refactor: jdk17 springboot3 sql init * refactor: jdk17 springboot3 sql init * Merge remote-tracking branch 'origin/dev' into dev * feat: course add * Merge remote-tracking branch 'origin/dev' into dev * feat: student login * Merge remote-tracking branch 'origin/dev' into dev * fix: responseError

+++ packages/backend/platform-service/src/main/java/com/teaching/platform/service/impl/CourseChapterServiceImpl.java
@@ -45,7 +45,7 @@ public class CourseChapterServiceImpl extends ServiceImpl<CourseChapterDao, Cour
     private List<CourseChapterResponse> extractedHasChild(List<CourseChapter> courseChapters) {
         List<CourseChapterResponse> courseChapterResponses = BeanHelper.castTo(courseChapters, CourseChapterResponse.class);
         for (CourseChapterResponse courseChapterResponse : courseChapterResponses) {
-            Integer count = this.baseMapper.selectCount(Wrappers.<CourseChapter>lambdaQuery().eq(CourseChapter::getParentId, courseChapterResponse.getParentId()));
+            Long count = this.baseMapper.selectCount(Wrappers.<CourseChapter>lambdaQuery().eq(CourseChapter::getParentId, courseChapterResponse.getParentId()));
             courseChapterResponse.setHasChild(MathHelper.nvl(count) > 0);
         }
         return courseChapterResponses;
